HSEARCH-2508 Revert "HSEARCH-2419 Split container bridges tests to test null-embedded naming conflicts separately"
This reverts commit c051bfe2ed8981482275ff63fbec3a5ea99d4cb3.

There is no point in splitting the tests anymore, and merging them again
reduces code duplication and makes fixing them easier.
