+++ orm/src/test/java/org/hibernate/search/test/bridge/MapBridgeTest.java
@@ -21,8 +21,10 @@ import org.hibernate.search.bridge.util.impl.NumericFieldUtils;
 import org.hibernate.search.query.dsl.QueryBuilder;
 import org.hibernate.search.query.dsl.TermMatchingContext;
 import org.hibernate.search.test.SearchTestBase;
+import org.hibernate.search.testsupport.junit.ElasticsearchSupportInProgress;
 import org.junit.Before;
 import org.junit.Test;
+import org.junit.experimental.categories.Category;
 
 import static org.hibernate.search.test.bridge.MapBridgeTestEntity.Language.ENGLISH;
 import static org.hibernate.search.test.bridge.MapBridgeTestEntity.Language.ITALIAN;
@@ -31,7 +33,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 /**
- * Test indexing of {@link javax.persistence.ElementCollection} annotated maps.
+ * Test indexing of {@link javax.persistence.ElementCollection} annotated elements.
  *
  * @author Davide D'Alto
  */
@@ -97,6 +99,27 @@ public class MapBridgeTest extends SearchTestBase {
 		assertNotNull( "No result found for an indexed collection", results );
 		assertEquals( "Unexpected number of results in a collection", 1, results.size() );
 		assertEquals( "Wrong result returned looking for a null in a collection", withNullEntry.getName(), results.get( 0 ).getName() );
+	}
+
+	@Test
+	@Category(ElasticsearchSupportInProgress.class) // HSEARCH-2389 Support indexNullAs for @IndexedEmbedded applied on objects with Elasticsearch
+	public void testSearchNullEmbedded() throws Exception {
+		List<MapBridgeTestEntity> results = findEmbeddedNullResults( "nullIndexed", MapBridgeTestEntity.NULL_EMBEDDED, true );
+
+		assertNotNull( "No result found for an indexed collection", results );
+		assertEquals( "Unexpected number of results in a collection", 1, results.size() );
+		assertEquals( "Wrong result returned looking for a null in a collection", withNullEmbedded.getName(), results.get( 0 ).getName() );
+	}
+
+	@Test
+	@Category(ElasticsearchSupportInProgress.class) // HSEARCH-2389 Support indexNullAs for @IndexedEmbedded applied on objects with Elasticsearch
+	public void testSearchNullNumericEmbedded() throws Exception {
+		List<MapBridgeTestEntity> results =
+				findEmbeddedNullResults( "embeddedNum", MapBridgeTestEntity.NULL_EMBEDDED_NUMERIC, true );
+
+		assertNotNull( "No result found for an indexed collection", results );
+		assertEquals( "Unexpected number of results in a collection", 1, results.size() );
+		assertEquals( "Wrong result returned looking for a null in a collection of numeric", withNullEmbedded.getName(), results.get( 0 ).getName() );
 	}
 
 	@Test
@@ -206,6 +229,18 @@ public class MapBridgeTest extends SearchTestBase {
 		assertEquals( "Wrong number of results returned for an indexed collection", 1, results.size() );
 		assertEquals( "Wrong result returned from a collection of Date", withoutNull.getName(), results.get( 0 )
 				.getName() );
+	}
+
+	@SuppressWarnings("unchecked")
+	private List<MapBridgeTestEntity> findEmbeddedNullResults(String fieldName, Object value, boolean checkNullToken) {
+		QueryBuilder queryBuilder = fullTextSession.getSearchFactory().buildQueryBuilder()
+				.forEntity( MapBridgeTestEntity.class ).get();
+		TermMatchingContext termMatchingContext = queryBuilder.keyword().onField( fieldName );
+		if ( checkNullToken ) {
+			termMatchingContext.ignoreFieldBridge();
+		}
+		Query query = termMatchingContext.ignoreAnalyzer().matching( value ).createQuery();
+		return fullTextSession.createFullTextQuery( query, MapBridgeTestEntity.class ).list();
 	}
 
 	@SuppressWarnings("unchecked")
